requirements.txt
Django==4.2
pandas
django-environ
djangorestframework
markdown
django-filter
djoser
social-auth-app-django

django-ckeditor

pytz

Pillow
mutagen 


#psycopg2

argon2-cffi
django-environ

stripe
braintree

pyJWT

boto3
django-storages 

markdown
gunicorn
whitenoise
Faker
requests

